requirements.txt
pystyle
pycryptodome
httpx